Fixed merging
The 'merge_table_with_results' method was failing in certain cases
(particularly when there were multiple <b>...</b> elements in the
string, resulting in a '!' in the diff context). Simplified the merge
method to include the json_results row in all cases except when the
original table had a row that json_results did not; also simplified
the 'clean_cell' method to just remove all <b>, </b> and <br> and <br/>
tags, instead of trying to pattern-match element content.