Multi-app test
The ability to test multiple projects is turning out to be much more
complicated than expected. Although it may not be too difficult to
enable it for AaaAccelerator-based test runs, there appears to be no
easy way to support specifying an AUT project directory in a way that
will be effective for individual table (do_table) tests, because those
are executed on a standalone basis (by passing the table's JSON text to
rubyslim), giving Cukable no awareness of what wiki-page-path the test
originated from. The only obvious way to get do_table to be aware of the
PROJECT_DIR argument is to pass it explicitly as an argument to the Cuke
constructor, i.e.:

    | Table: Cuke | ${PROJECT_DIR} |

While this would be a workable solution, it is an egregious violation of
the DRY principle. I see this as a deal-breaker for multi-project
support in Cukable.

It may be worth investigating whether RubySlim, as the intermediary
between FitNesse and Cukable, could be hacked to provide a more
satisfactory solution. For now, I'm not entirely convinced that
single-app testing with Cukable is viable, let alone multi-app testing.

Setting this branch aside for now.