Require `features` so step definitions can be found
The cucumber command, by virtue of using the `--require` option to
include the SlimJSON formatter, was not correctly including step
definitions that could be defined in sibling directories of the
features/ directory. It's necessary to explicitly include features/ to
get these included. Turns out one of the self-tests was wrong (expecting
error, meaning undefined step, instead of pass/fail like it should have
had). Fixed.